chore(jaeger): Move to be an internal package.
